face_search/requirements.txt
numpy
pillow
mtcnn
keras-facenet
faiss-cpu
scipy